뷰 홀더, DiffCallback 어댑터에서 분리
